Create displayUI class for handeling puts and gets
